+++ Core/Types/CkanModule.cs
@@ -140,6 +140,16 @@ namespace CKAN
         [JsonProperty("install", Order = 24, NullValueHandling = NullValueHandling.Ignore)]
         public ModuleInstallDescriptor[]? install;
 
+        public ModuleInstallDescriptor[] GetInstallStanzas(IGame game)
+            => install switch
+               {
+                   { Length: > 0 } => install,
+                   _ => new ModuleInstallDescriptor[]
+                        {
+                            ModuleInstallDescriptor.DefaultInstallStanza(game, identifier)
+                        }
+               };
+
         [JsonProperty("localizations", Order = 17, NullValueHandling = NullValueHandling.Ignore)]
         public string[]? localizations;
 
@@ -182,23 +192,14 @@ namespace CKAN
         [JsonProperty("release_date", Order = 30, NullValueHandling = NullValueHandling.Ignore)]
         public DateTime? release_date;
 
+        [JsonIgnore]
+        private string[]? providesList = null;
+
         // A list of eveything this mod provides.
-        public List<string> ProvidesList
-        {
-            // TODO: Consider caching this, but not in a way that the serialiser will try and
-            // serialise it.
-            get
-            {
-                var provides = new List<string> { identifier };
-
-                if (this.provides != null)
-                {
-                    provides.AddRange(this.provides);
-                }
-
-                return provides;
-            }
-        }
+        public string[] ProvidesList
+            => providesList ??= (provides ?? Enumerable.Empty<string>())
+                                    .Prepend(identifier)
+                                    .ToArray();
 
         // These are used to simplify the search by dropping special chars.
         [JsonIgnore]
@@ -714,10 +715,7 @@ namespace CKAN
             => string.Format("{0} {1}", identifier, version);
 
         public string DescribeInstallStanzas(IGame game)
-            => install == null
-                ? ModuleInstallDescriptor.DefaultInstallStanza(game, identifier)
-                                         .DescribeMatch()
-                : string.Join(", ", install.Select(mid => mid.DescribeMatch()));
+            => string.Join(", ", GetInstallStanzas(game).Select(mid => mid.DescribeMatch()));
 
         /// <summary>
         /// Return an archive.org URL for this download, or null if it's not there.